+++ lib/src/ast/selector/list.dart
@@ -36,15 +36,34 @@ final class SelectorList extends Selector {
   SassList get asSassList {
     return SassList(components.map((complex) {
       return SassList([
-        for (var combinator in complex.leadingCombinators)
+        if (complex.leadingCombinator case var combinator?)
           SassString(combinator.toString(), quotes: false),
         for (var component in complex.components) ...[
           SassString(component.selector.toString(), quotes: false),
-          for (var combinator in component.combinators)
+          if (component.combinator case var combinator?)
             SassString(combinator.toString(), quotes: false)
         ]
       ], ListSeparator.space);
     }), ListSeparator.comma);
+  }
+
+  /// Whether this selector is invalid CSS that's useful exclusively for
+  /// build-time nesting (`> .foo)`.
+  bool get isBogus => components.any((complex) => complex.isBogus);
+
+  /// Whether this selector is bogus other than having a leading combinator.
+  ///
+  /// @nodoc
+  @internal
+  bool get isBogusOtherThanLeadingCombinator =>
+      components.any((complex) => complex.isBogusOtherThanLeadingCombinator);
+
+  /// Prints a warning if `this` is a bogus selector.
+  ///
+  /// This may only be called from within a custom Sass function. This will
+  /// throw a [SassException] in Dart Sass 2.0.0.
+  void assertNotBogus({String? name}) {
+    if (isBogus) throw SassException('Invalid CSS.', span);
   }
 
   SelectorList(Iterable<ComplexSelector> components, super.span)
@@ -121,8 +140,16 @@ final class SelectorList extends Selector {
     return SelectorList(flattenVertically(components.map((complex) {
       if (preserveParentSelectors || !_containsParentSelector(complex)) {
         if (!implicitParent) return [complex];
-        return parent.components.map((parentComplex) =>
-            parentComplex.concatenate(complex, complex.span));
+        return [
+          for (var parentComplex in parent.components)
+            if (parentComplex.concatenate(complex, complex.span)
+                case var newComplex?)
+              newComplex
+            else
+              throw SassException(
+                  'The selector "$parentComplex $complex" is invalid CSS.',
+                  complex.span)
+        ];
       }
 
       var newComplexes = <ComplexSelector>[];
@@ -130,8 +157,8 @@ final class SelectorList extends Selector {
         var resolved = _nestWithinCompound(component, parent);
         if (resolved == null) {
           if (newComplexes.isEmpty) {
-            newComplexes.add(ComplexSelector(
-                complex.leadingCombinators, [component], complex.span,
+            newComplexes.add(ComplexSelector([component], complex.span,
+                leadingCombinator: complex.leadingCombinator,
                 lineBreak: false));
           } else {
             for (var i = 0; i < newComplexes.length; i++) {
@@ -140,24 +167,32 @@ final class SelectorList extends Selector {
             }
           }
         } else if (newComplexes.isEmpty) {
-          newComplexes.addAll(complex.leadingCombinators.isEmpty
-              ? resolved
-              : resolved.map((resolvedComplex) => ComplexSelector(
-                  resolvedComplex.leadingCombinators.isEmpty
-                      ? complex.leadingCombinators
-                      : [
-                          ...complex.leadingCombinators,
-                          ...resolvedComplex.leadingCombinators
-                        ],
-                  resolvedComplex.components,
-                  complex.span,
-                  lineBreak: resolvedComplex.lineBreak)));
+          newComplexes.addAll(switch (complex.leadingCombinator) {
+            null => resolved,
+            var leadingCombinator? => [
+                for (var resolvedComplex in resolved)
+                  if (resolvedComplex.prependCombinator(leadingCombinator)
+                      case var newResolved?)
+                    newResolved
+                  else
+                    throw SassException(
+                        'The selector "$leadingCombinator $resolvedComplex" is '
+                        'invalid CSS.',
+                        complex.span)
+              ]
+          });
         } else {
-          var previousComplexes = newComplexes;
           newComplexes = [
-            for (var newComplex in previousComplexes)
+            for (var newComplex in newComplexes)
               for (var resolvedComplex in resolved)
-                newComplex.concatenate(resolvedComplex, newComplex.span)
+                if (newComplex.concatenate(resolvedComplex, newComplex.span)
+                    case var newResolved?)
+                  newResolved
+                else
+                  throw SassException(
+                      'The selector "$newComplex $resolvedComplex" is invalid '
+                      'CSS.',
+                      complex.span)
           ];
         }
       }
@@ -183,30 +218,39 @@ final class SelectorList extends Selector {
     }
 
     var resolvedSimples = containsSelectorPseudo
-        ? simples.map((simple) => switch (simple) {
-              PseudoSelector(:var selector?)
-                  when _containsParentSelector(selector) =>
-                simple.withSelector(
-                    selector.nestWithin(parent, implicitParent: false)),
-              _ => simple
-            })
+        ? simples.map((simple) {
+            if (simple case PseudoSelector(:var selector?)
+                when _containsParentSelector(selector)) {
+              var nested = selector.nestWithin(parent, implicitParent: false);
+              var result = simple.withSelector(nested);
+              if (result != null) return result;
+
+              var invalid = simple
+                  .toString()
+                  .replaceFirst(RegExp(r"\(.*\)"), "($nested)");
+              throw SassException(
+                  'The selector "$invalid" is invalid CSS.', simple.span);
+            } else {
+              return simple;
+            }
+          })
         : simples;
 
     var parentSelector = simples.first;
     try {
       if (parentSelector is! ParentSelector) {
         return [
-          ComplexSelector(const [], [
+          ComplexSelector([
             ComplexSelectorComponent(
                 CompoundSelector(resolvedSimples, component.selector.span),
-                component.combinators,
-                component.span)
+                component.span,
+                combinator: component.combinator)
           ], component.span)
         ];
       } else if (simples.length == 1 && parentSelector.suffix == null) {
         return parent
-            .withAdditionalCombinators(component.combinators)
-            .components;
+            .withAdditionalCombinator(component.combinator)
+            ?.components;
       }
     } on SassException catch (error, stackTrace) {
       throwWithTrace(
@@ -218,7 +262,7 @@ final class SelectorList extends Selector {
     return parent.components.map((complex) {
       try {
         var lastComponent = complex.components.last;
-        if (lastComponent.combinators.isNotEmpty) {
+        if (lastComponent.combinator != null) {
           throw MultiSpanSassException(
               'Selector "$complex" can\'t be used as a parent in a compound '
                   'selector.',
@@ -239,14 +283,12 @@ final class SelectorList extends Selector {
                   ],
             component.selector.span);
 
-        return ComplexSelector(
-            complex.leadingCombinators,
-            [
-              ...complex.components.exceptLast,
-              ComplexSelectorComponent(
-                  last, component.combinators, component.span)
-            ],
-            component.span,
+        return ComplexSelector([
+          ...complex.components.exceptLast,
+          ComplexSelectorComponent(last, component.span,
+              combinator: component.combinator)
+        ], component.span,
+            leadingCombinator: complex.leadingCombinator,
             lineBreak: complex.lineBreak);
       } on SassException catch (error, stackTrace) {
         throwWithTrace(
@@ -264,19 +306,22 @@ final class SelectorList extends Selector {
   bool isSuperselector(SelectorList other) =>
       listIsSuperselector(components, other.components);
 
-  /// Returns a copy of `this` with [combinators] added to the end of each
+  /// Returns a copy of `this` with [combinator] added to the end of each
   /// complex selector in [components].
+  ///
+  /// Returns `null` if this would produce an invalid selector.
   ///
   /// @nodoc
   @internal
-  SelectorList withAdditionalCombinators(
-          List<CssValue<Combinator>> combinators) =>
-      combinators.isEmpty
-          ? this
-          : SelectorList(
-              components.map(
-                  (complex) => complex.withAdditionalCombinators(combinators)),
-              span);
+  SelectorList? withAdditionalCombinator(CssValue<Combinator>? combinator) {
+    if (combinator == null) return this;
+    var newComponents = [
+      for (var complex in components)
+        if (complex.withAdditionalCombinator(combinator) case var newComplex?)
+          newComplex
+    ];
+    return newComponents.isEmpty ? null : SelectorList(newComponents, span);
+  }
 
   int get hashCode => listHash(components);